[Tabs] Decouple strict parent/child hierarchy
For TabList/Tab and TabPanels/TabPanel, the implementation before this
commit would require those components to have a strict parent/child
relationship. By using a "[context registering]" pattern, we relieve
this tight coupling.

This decoupling change provides the flexibility for other components to
be interwoven with the Tabs components, as can be seen in the with-wrappers
Storybook example file.

NOTE: this pattern will likely break in some cases when React is running
in [Concurrent Mode]. Since each Tab and TabPanel will register itself in
the order it is rendered, if some tabs or panels are wrapped in the
[Suspense] primitive they may register out of order, which could result
in their data index differing from their visual order.

[context registering]: https://gist.github.com/ryanflorence/e5c794e6093d16a69fa88d2112a292f7#additional-considerations
[Concurrent Mode]: https://reactjs.org/docs/concurrent-mode-intro.html
[Suspense]: https://reactjs.org/docs/concurrent-mode-patterns.html#wrap-lazy-features-in-suspense